+++ .gitignore
@@ -20,6 +20,7 @@
 __pycache__
 /cmd/podman/varlink/ioprojectatomicpodman.go
 /cmd/podman/varlink/iopodman.go
+/pkg/varlink/iopodman.go
 .gopathok
 release.txt
 podman-remote*.zip
