+++ cmd/podman/errors_remote.go
@@ -8,8 +8,8 @@ import (
 	"os/exec"
 	"syscall"
 
-	"github.com/containers/libpod/cmd/podman/varlink"
 	"github.com/containers/libpod/libpod/define"
+	iopodman "github.com/containers/libpod/pkg/varlink"
 	"github.com/pkg/errors"
 	"github.com/sirupsen/logrus"
 )
