+++ cmd/podman/shared/intermediate_varlink.go
@@ -4,8 +4,8 @@ package shared
 
 import (
 	"github.com/containers/libpod/cmd/podman/cliconfig"
-	"github.com/containers/libpod/cmd/podman/varlink"
 	"github.com/containers/libpod/pkg/rootless"
+	iopodman "github.com/containers/libpod/pkg/varlink"
 	"github.com/pkg/errors"
 )
